userbot/requirements.txt
aiogram==2.25.1
requests[all]
pydantic[all]
python-dotenv